ADDED: Test case to get invoice summary using user id

+++ src/main/java/com/bridgelabz/invoicegenerator/InvoiceGenerator.java
@@ -2,9 +2,9 @@ package com.bridgelabz.invoicegenerator;
 
 public class InvoiceGenerator {
 
-	private static final double COST_PER_KILOMETER = 10;
-	private static final int COST_PER_TIME = 1;
-	private static final double MINIMUM_FARE = 5;
+	private static final double COST_PER_KILOMETER = 10.0;
+	private static final double COST_PER_TIME = 1.0;
+	private static final double MINIMUM_FARE = 5.0;
 
 	public double calculateFare(double distance, int time) {
 		double totalFare = distance * COST_PER_KILOMETER + time * COST_PER_TIME;
@@ -14,11 +14,11 @@ public class InvoiceGenerator {
 
 	public InvoiceSummary calculateFare(Ride[] rides) {
 		double totalFare = 0;
-		for(Ride ride : rides) {
+		for (Ride ride : rides) {
 			double fare = this.calculateFare(ride.getDistance(), ride.getTime());
 			totalFare += fare;
 		}
-		return new InvoiceSummary(rides.length,totalFare);
+		return new InvoiceSummary(rides.length, totalFare);
 	}
 
 }
